Finish stage environment updates
